Revert "[JAVA-873] Extracted benchmarking part of unit test into its own LiveTest"
This reverts commit 10c05ebb7e43be0a04ed377c453f0aa92c81117c.
