+++ core-java-modules/core-java-concurrency-advanced-3/src/test/java/com/baeldung/workstealing/PrimeNumbersUnitTest.java
@@ -1,6 +1,7 @@
 package com.baeldung.workstealing;
 
 import org.junit.Test;
+import org.openjdk.jmh.annotations.*;
 import org.openjdk.jmh.runner.Runner;
 import org.openjdk.jmh.runner.RunnerException;
 import org.openjdk.jmh.runner.options.Options;
@@ -8,6 +9,7 @@ import org.openjdk.jmh.runner.options.OptionsBuilder;
 
 import java.util.concurrent.Executors;
 import java.util.concurrent.ForkJoinPool;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.logging.Logger;
 
@@ -16,6 +18,20 @@ import static org.junit.Assert.fail;
 public class PrimeNumbersUnitTest {
 
     private static Logger logger = Logger.getAnonymousLogger();
+
+    @Test
+    public void givenPrimesCalculated_whenUsingPoolsAndOneThread_thenOneThreadSlowest() {
+        Options opt = new OptionsBuilder()
+          .include(Benchmarker.class.getSimpleName())
+          .forks(1)
+          .build();
+
+        try {
+            new Runner(opt).run();
+        } catch (RunnerException e) {
+            fail();
+        }
+    }
 
     @Test
     public void givenNewWorkStealingPool_whenGettingPrimes_thenStealCountChanges() {
@@ -50,5 +66,36 @@ public class PrimeNumbersUnitTest {
         long steals = forkJoinPool.getStealCount();
         String output = "\nGranularity: [" + granularity + "], Steals: [" + steals + "]";
         info.append(output);
+    }
+
+
+    @BenchmarkMode(Mode.AverageTime)
+    @OutputTimeUnit(TimeUnit.MILLISECONDS)
+    @State(Scope.Benchmark)
+    @Fork(value = 2, warmups = 1, jvmArgs = {"-Xms2G", "-Xmx2G"})
+    public static class Benchmarker {
+
+        @Benchmark
+        public void singleThread() {
+            PrimeNumbers primes = new PrimeNumbers(10000);
+            primes.findPrimeNumbers(); // get prime numbers using a single thread
+        }
+
+        @Benchmark
+        public void commonPoolBenchmark() {
+            PrimeNumbers primes = new PrimeNumbers(10000);
+            ForkJoinPool pool = ForkJoinPool.commonPool();
+            pool.invoke(primes);
+            pool.shutdown();
+        }
+
+        @Benchmark
+        public void newWorkStealingPoolBenchmark() {
+            PrimeNumbers primes = new PrimeNumbers(10000);
+            int parallelism = ForkJoinPool.getCommonPoolParallelism();
+            ForkJoinPool stealer = (ForkJoinPool) Executors.newWorkStealingPool(parallelism);
+            stealer.invoke(primes);
+            stealer.shutdown();
+        }
     }
 }
